+++ selfdrive/ui/replay/logreader.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <unordered_map>
-#include <vector>
 #include <cassert>
 
 #include <capnp/serialize.h>
@@ -18,36 +17,7 @@ public:
     this->which = which;
     this->mono_time = mono_time;
   }
-  Event(const kj::ArrayPtr<const capnp::word> &amsg, bool frame=false) : reader(amsg), frame(frame) {
-    words = kj::ArrayPtr<const capnp::word>(amsg.begin(), reader.getEnd());
-    event = reader.getRoot<cereal::Event>();
-    which = event.which();
-    mono_time = event.getLogMonoTime();
-
-    // 1) Send video data at t=timestampEof/timestampSof
-    // 2) Send encodeIndex packet at t=logMonoTime
-    if (frame) {
-      cereal::EncodeIndex::Reader idx;
-      if (which == cereal::Event::ROAD_ENCODE_IDX) {
-        idx = event.getRoadEncodeIdx();
-      } else if (which == cereal::Event::DRIVER_ENCODE_IDX) {
-        idx = event.getDriverEncodeIdx();
-      } else if (which == cereal::Event::WIDE_ROAD_ENCODE_IDX) {
-        idx = event.getWideRoadEncodeIdx();
-      } else {
-        assert(false);
-      }
-
-      // C2 only has eof set, and some older routes have neither
-      uint64_t sof = idx.getTimestampSof();
-      uint64_t eof = idx.getTimestampEof();
-      if (sof > 0) {
-        mono_time = sof;
-      } else if (eof > 0) {
-        mono_time = eof;
-      }
-    }
-  }
+  Event(const kj::ArrayPtr<const capnp::word> &amsg, bool frame = false);
   inline kj::ArrayPtr<const capnp::byte> bytes() const { return words.asBytes(); }
 
   struct lessThan {
@@ -68,10 +38,10 @@ class LogReader {
 public:
   LogReader() = default;
   ~LogReader();
-  bool load(const std::string &file);
+  bool load(const std::string &file, bool is_bz2file);
 
   std::vector<Event*> events;
 
 private:
-  std::vector<uint8_t> raw_;
+  std::string raw_;
 };
