+++ selfdrive/ui/replay/route.h
@@ -67,5 +67,3 @@ protected:
   QString log_path_;
   std::vector<QThread*> download_threads_;
 };
-
-bool httpMultiPartDownload(const std::string &url, const std::string &target_file, int parts, std::atomic<bool> *abort = nullptr);
